requirements.txt
Appium-Python-Client==5.2.7
attrs==25.4.0
beautifulsoup4==4.14.3
behave==1.3.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
cryptography==46.0.5
cucumber-expressions==19.0.0
cucumber-tag-expressions==9.1.0
filelock==3.25.2
gdown==5.2.1
google-api-core==2.30.0
google-auth==2.49.1
googleapis-common-protos==1.73.0
h11==0.16.0
httplib2==0.31.2
idna==3.11
oauthlib==3.3.1
outcome==1.3.0.post0
parse==1.21.1
parse_type==0.6.6
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
PySocks==1.7.1
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
selenium==4.41.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
